// portal/customer-portal-vivo-angular/src/app/enumerators/phone-type.enum.ts
export enum PhoneType {

    // DECRIÇÃO DOS ENUMERADORES
    PHONE_CELULAR = 'Celular',
    PHONE_TELEFONE_CONTATO = 'Telefone de Contato',
    PHONE_TELEFONE_TRABALHO = 'Telefone Comercial',
    PHONE_TELEFONE_RESIDENCIA = 'Telefone Residencial',

    // IDENTIFICAÇÃO DOS ENUMERADORES
    PHONE_CELULAR_ID = 'Mobile',
    PHONE_TELEFONE_CONTATO_ID = 'Contact',
    PHONE_TELEFONE_TRABALHO_ID = 'Work',
    PHONE_TELEFONE_RESIDENCIA_ID = 'Home',
}