// portal/customer-portal-vivo-angular/src/app/enumerators/road-type.enum.ts
export enum RoadType {

    // DECRIÇÃO DOS ENUMERADORES
    TIPO_LOGRADOURO_RUA = 'Rua',
    TIPO_LOGRADOURO_AVENIDA = 'Avenida',
    TIPO_LOGRADOURO_VIELA = 'Viela',
    TIPO_LOGRADOURO_ESQUINA = 'Esquina',
    TIPO_LOGRADOURO_PERIMETRAL = 'Perimetral',
    TIPO_LOGRADOURO_PRACA = 'Praça',
    TIPO_LOGRADOURO_TREVO = 'Trevo',
    TIPO_LOGRADOURO_ROTATORIA = 'Rotatoria',


    // IDENTIFICAÇÃO DOS ENUMERADORES
    TIPO_LOGRADOURO_RUA_ID = 'Rua',
    TIPO_LOGRADOURO_AVENIDA_ID = 'Avenida',
    TIPO_LOGRADOURO_VIELA_ID = 'Viela',
    TIPO_LOGRADOURO_ESQUINA_ID = 'Esquina',
    TIPO_LOGRADOURO_PERIMETRAL_ID = 'Perimetral',
    TIPO_LOGRADOURO_PRACA_ID = 'Praça',
    TIPO_LOGRADOURO_TREVO_ID = 'Trevo',
    TIPO_LOGRADOURO_ROTATORIA_ID = 'Rotatoria',


}